bugs fixed V2
