+++ models/user.js
@@ -1,6 +1,8 @@
 const mongoose = require('mongoose');
 const bcrypt = require('bcrypt'); // импортируем bcrypt
 
+const { AuthError } = require('../errors/AuthError');
+
 const userSchema = new mongoose.Schema({
   name: {
     type: String,
@@ -31,12 +33,12 @@ userSchema.statics.findUserByCredentials = function (email, password) {
     .select('+password') // из-за запрета возврата хеш пароля
     .then((user) => {
       if (!user) {
-        return Promise.reject(new Error('Неправильные почта или пароль'));
+        return Promise.reject(new AuthError('Неправильные почта или пароль'));
       }
 
       return bcrypt.compare(password, user.password).then((matched) => {
         if (!matched) {
-          return Promise.reject(new Error('Неправильные почта или пароль'));
+          return Promise.reject(new AuthError('Неправильные почта или пароль'));
         }
 
         return user;
