+++ utils/validators.js
@@ -23,7 +23,7 @@ const signInValidator = celebrate({
 const changeProfileValidator = celebrate({
   body: Joi.object().keys({
     name: Joi.string().min(2).max(30).required(),
-    email: Joi.string().email(),
+    email: Joi.string().email().required(),
   }),
 });
